+++ ToreadorRESTAPI/src/main/java/uk/ac/city/toreador/rest/api/jpa/repositories/GuardedActionsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import uk.ac.city.toreador.rest.api.entities.GuardedAction;
 
-public interface GuardedActionsRepository extends CrudRepository<GuardedAction, Long>{
+public interface GuardedActionsRepository extends CrudRepository<GuardedAction, Integer>{
 	
 }
